refactor(stores): split demo store into domain-specific sub-stores
The demo store has been refactored from a monolithic 500-line file into
a modular directory structure with domain-specific slices:

- types.ts: Shared type definitions and constants
- assignments.ts: Assignment data state
- compensations.ts: Compensation records and edits
- exchanges.ts: Game exchange operations
- nominations.ts: Nomination lists and players
- validation.ts: Game validation and pending scorers
- index.ts: Combined store with lifecycle actions

This improves maintainability by isolating domain logic, making each
slice easier to test and modify independently. The original demo.ts
file now re-exports from the new structure for backward compatibility.

Closes #281
